requirements.txt
pymongo
tkinter
python-dotenv